! First success in getting data from the database
This is the very first successfull attempt in getting the data from
firebird database in about five days fault. The reason of whole bunch
of the mistakes was the 32 bit version of Python is needed. Once the
version has been installed the connection to the base and retrieving
the data started working properly. The code is in viwes.py